Fix compilation on macOS + clang-14
1. Added missing prototypes to the global header.
2. Removed invalid escape sequences.
3. Fixed signed vs unsigned comparisons.
4. Fixed duplicated BUFSIZ definition.
5. Added needed flags to the clang makefile.

+++ parser.c
@@ -4142,7 +4142,7 @@ void ProcessFile( struct asm_tok tokenarray[] )
 
     if ( ModuleInfo.EndDirFound == FALSE && GetTextLine( CurrSource ) ) 
 	{
-		if (CurrSource[0] == 0xEF && CurrSource[1] == 0xBB && CurrSource[2] == 0xBF)
+		if ((uint8_t)CurrSource[0] == 0xEF && (uint8_t)CurrSource[1] == 0xBB && (uint8_t)CurrSource[2] == 0xBF)
 			strcpy(CurrSource, &CurrSource[3]);
 		do
 		{
